+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 5.14.0
-* #1166 add method `SelenideDriver.screenshot(fileName)`
+* #1194 added method `$$.shouldHave(itemWithText("any text"))`  --  thanks to Luis Serna for PR #1194
+* #1166 added method `SelenideDriver.screenshot(fileName)`
 * upgraded to WebDriverManager 4.1.0
 
 ## 5.13.0 (released 08.07.2020)